double, mostly done

+++ src/libsurskit/math.cc
@@ -1,10 +1,10 @@
 #include "math.hh"
 
 // equivalent to multiplying the logits by 'power'
-void math::power_norm(float* input, int k, float power, float* output) {
-    float sum = 0;
+void math::power_norm(double* input, int k, double power, double* output) {
+    double sum = 0;
     for (int i = 0; i < k; ++i) {
-        float x = std::pow(input[i], power);
+        double x = std::pow(input[i], power);
         output[i] = x;
         sum += x;
     }
